Bank on MSIL
While the X64 backend was fun to work on it took a really long time to get
anywhere. While that was going on I started thinking "Do I need the potential
freedom and performance gains by doing this?".
While I can't answer that for sure I thought that Underrail runs on C# and
MonoGame and my current target is to do what I immagine is a simpler
turn based game using Owen and SDL2 so MSIL is probably more
than enough. If not I could dust off the X64 backend bits I have lying around
and give it another go or deep dive into LLVM.